% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringLiger.R
\name{clusteringLiger}
\alias{clusteringLiger}
\title{LIGER clustering using iNMF or UINMF algorythm}
\usage{
clusteringLiger(
  datasets,
  lambdaValue = 5,
  kValue = 25,
  referenceDatasetName,
  resolutionValue = 0.4,
  setVariableGenes = FALSE,
  num.genesValue = NULL,
  var.threshValue = NULL,
  useiNMF = FALSE,
  useUINMF = FALSE,
  unshared.threshValue = NULL
)
}
\arguments{
\item{datasets}{list of named datasets}

\item{lambdaValue}{default 5}

\item{kValue}{default 25}

\item{referenceDatasetName}{name of dataset to use as reference dataset}

\item{resolutionValue}{default 0.4}

\item{setVariableGenes}{default FALSE}

\item{num.genesValue}{default NULL}

\item{var.threshValue}{default NULL}

\item{useiNMF}{default FALSE}

\item{useUINMF}{default FALSE}

\item{unshared.threshValue}{default NULL}
}
\description{
Function that takes a named list of datasets and integrates using LIGER
}
