% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringSCT.R
\name{clusteringSCT}
\alias{clusteringSCT}
\title{Seurat clustering with SCT transformation}
\usage{
clusteringSCT(
  datasetObject,
  selectFeatures = NULL,
  selectedFeatureNames,
  pcaValueOptimal = NULL,
  datasetName,
  resolutionValue = 0.2
)
}
\arguments{
\item{datasetObject}{dataset object}

\item{selectFeatures}{choose if features are to be selected (default NULL)}

\item{selectedFeatureNames}{list of features to use on the clustering}

\item{pcaValueOptimal}{define a specific PCA value to use (default NULL). Otherwise it will calculate an optimal PCA value.}

\item{datasetName}{define the name of the dataset}

\item{resolutionValue}{define the resolution value to use for clustering (default 0.2)}

\item{metadataObject}{metadata object}
}
\description{
Function that takes a dataset (either spatial or scRNA-seq), creates Seurat object, runs SCT normalization, PCA, etc, clustering
}
