% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSpatialSeurat.R
\name{createSpatialSeurat}
\alias{createSpatialSeurat}
\title{Seurat spatial object}
\usage{
createSpatialSeurat(datasetObject, metadataObject)
}
\arguments{
\item{datasetObject}{dataset object}

\item{metadataObject}{metadata object that includes X and Y coordinates for each cell}
}
\description{
Function that takes a spatial dataset and respective x and y coordinates to generate the spatial map of an individual sample
}
