% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringSeurat.R
\name{clusteringSeurat}
\alias{clusteringSeurat}
\title{Seurat clustering with log or SCT normalization}
\usage{
clusteringSeurat(
  datasetObject,
  datasetName,
  metadataAvailable = TRUE,
  normalizationMethod = "LOG",
  mapTypeValue = "umap",
  metadataObject,
  reductionValue = "umap",
  resolutionValue = 0.1,
  pcaValueOptimal = NULL,
  integrateValue = FALSE,
  specificPCA = FALSE,
  ngenes = 200,
  scaleFactor = 10000,
  nPCAS = 30,
  selectFeatures = NULL
)
}
\arguments{
\item{datasetObject}{dataset object}

\item{datasetName}{name of the dataset}

\item{metadataAvailable}{whether there is prior metadata available (default TRUE)}

\item{normalizationMethod}{which normalization method to use (default LOG). options log normalization or SCTransform normalization (SCT)}

\item{metadataObject}{metadata object}

\item{reductionValue}{which dimensionality reduction to use, umap or tsne (default umap)}

\item{resolutionValue}{define the resolution value to use for clustering (default 0.1)}

\item{pcaValueOptimal}{define a specific PCA value to use (default NULL calculates automatically the optimal PCA value)}

\item{integrateValue}{whether dataset integration is needed (default FALSE)}

\item{specificPCA}{whether PCA values needs to be calculated or not}

\item{ngenes}{minimum number of features expressed in a cell (default 200)}

\item{scaleFactor}{value for scaling (default 10000)}

\item{nPCAS}{number of principal components to use (default 30)}

\item{selectFeatures}{whether features need to be selected. Takes all features as default (default NULL)}
}
\description{
Function that takes a dataset (either spatial or scRNA-seq), creates Seurat object, runs log normalization, PCA, etc, clustering
}
